move jsx files
