+++ .gitignore
@@ -41,6 +41,7 @@
 
 config/local.js
 .vscode/
+.sass-cache/
