--- assets/app/main.js
@@ -1,12 +1,0 @@
-var ReactDOM = require('react-dom');
-var React = require('react');
-
-// var path = require('path');
-
-// var footerPath = path.resolve('../../') + '/views/Footer.jsx';
-
-var AppRouter = require('/home/craig/Documents/sites/ratemybill-site/views/AppRouter.jsx');
-// var Footer = require(footerPath);
-
-ReactDOM.render(<AppRouter />, document.getElementById('app'));
-// ReactDOm.render(<Footer />, document.getElementById('footer'));
