+++ assets/app/src/components/AboutPage.jsx
@@ -1,4 +1,4 @@
-var React = require('react');
+import React from 'react';
 
 var AboutPage = React.createClass({
     render: function() {
