+++ assets/app/src/components/Footer.jsx
