+++ assets/styles/styles.scss
@@ -1,0 +1,4 @@
+@charset "UTF-8";
+
+@import "bower_components/materialize/sass/materialize.scss";
+@import "./styles.scss";
